Linux support for patcher.

+++ patcher_src/main.go
@@ -6,6 +6,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"runtime"
 )
 
 func main() {
@@ -15,41 +16,43 @@ func main() {
 	}
 	exe_dir := filepath.Dir(exe)
 
-	out, err := exec.Command(exe_dir + "/utils/setdll64.exe", "/d:portable64.dll", exe_dir + "/core/mozglue.dll").Output()
-	fmt.Println(string(out))
-	if err != nil {
-		panic(err)
-	}
+	if runtime.GOOS == "windows" {
+		out, err := exec.Command(exe_dir + "/utils/setdll64.exe", "/d:portable64.dll", exe_dir + "/core/mozglue.dll").Output()
+		fmt.Println(string(out))
+		if err != nil {
+			panic(err)
+		}
 
-	src, err := os.Open(exe_dir + "/utils/portable64.dll")
-	if err != nil {
-		panic(err)
-	}
-	dst, err := os.Create(exe_dir + "/core/portable64.dll")
-	if err != nil {
-		panic(err)
-	}
-	_, err = io.Copy(dst, src)
-	if  err != nil {
-		panic(err)
-	}
-	src.Close()
-	dst.Close()
+		src, err := os.Open(exe_dir + "/utils/portable64.dll")
+		if err != nil {
+			panic(err)
+		}
+		dst, err := os.Create(exe_dir + "/core/portable64.dll")
+		if err != nil {
+			panic(err)
+		}
+		_, err = io.Copy(dst, src)
+		if  err != nil {
+			panic(err)
+		}
+		src.Close()
+		dst.Close()
 
-	src, err = os.Open(exe_dir + "/config/portable.ini")
-	if err != nil {
-		panic(err)
+		src, err = os.Open(exe_dir + "/config/portable.ini")
+		if err != nil {
+			panic(err)
+		}
+		dst, err = os.Create(exe_dir + "/core/portable.ini")
+		if err != nil {
+			panic(err)
+		}
+		_, err = io.Copy(dst, src)
+		if  err != nil {
+			panic(err)
+		}
+		src.Close()
+		dst.Close()
 	}
-	dst, err = os.Create(exe_dir + "/core/portable.ini")
-	if err != nil {
-		panic(err)
-	}
-	_, err = io.Copy(dst, src)
-	if  err != nil {
-		panic(err)
-	}
-	src.Close()
-	dst.Close()
 
 	if _, err := os.Stat(exe_dir + "/core/distribution"); err != nil {
 		err := os.Mkdir(exe_dir + "/core/distribution", 0777)
@@ -58,11 +61,11 @@ func main() {
 		}
 	}
 
-	src, err = os.Open(exe_dir + "/config/policies.json")
+	src, err := os.Open(exe_dir + "/config/policies.json")
 	if err != nil {
 		panic(err)
 	}
-	dst, err = os.Create(exe_dir + "/core/distribution/policies.json")
+	dst, err := os.Create(exe_dir + "/core/distribution/policies.json")
 	if err != nil {
 		panic(err)
 	}
@@ -88,16 +91,18 @@ func main() {
 	src.Close()
 	dst.Close()
 
-	err = os.Remove(exe_dir + "/core/updater.exe")
-	if  err != nil {
-		panic(err)
-	}
-	err = os.Remove(exe_dir + "/core/default-browser-agent.exe")
-	if  err != nil {
-		panic(err)
-	}
-	err = os.RemoveAll(exe_dir + "/core/uninstall")
-	if  err != nil {
-		panic(err)
+	if runtime.GOOS == "windows" {
+		err := os.Remove(exe_dir + "/core/updater.exe")
+		if  err != nil {
+			panic(err)
+		}
+		err = os.Remove(exe_dir + "/core/default-browser-agent.exe")
+		if  err != nil {
+			panic(err)
+		}
+		err = os.RemoveAll(exe_dir + "/core/uninstall")
+		if  err != nil {
+			panic(err)
+		}
 	}
 }
